requirements.txt
Flask>=2.3
Flask-Login
Flask-SQLAlchemy
pymysql
python-dotenv
pandas
weasyprint
Werkzeug
